backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-jose
passlib
bcrypt
python-multipart
python-dotenv
email-validator
